Remove checkbox again
Too much overhead for too little use. Instead, a todo's done status can
now simiply be toggled by right-clicking it while outside the edit mode.
Todos that are done will be modulated to 30% of their alpha value. If a
todo is marked as done, it cannot be moved via drag and drop anymore. An
empty panel will be used by default now.