Implement rudimentary save system
If a DayView scene is removed from the tree due to the user using the
navigation bar, its content (minus debug todos) is written to disk and
restored when the respective day is loaded again later on.

If there already is a file, but no content anymore (i.e. the user edited
the existing todos to an empty string) the file gets deleted.