Save after each edit
Each todo item now emits a `editing_finished` signal that prompts the
todo list holding it to save to disk. The signal will *not* be emitted
when submitting an empty todo or headline. However, for now, it is still
emitted when a todo enters and exits edit mode without a change - will
fix that in a future commit, since right now I don't see a simple and
straightforward solution to that.

Also, the todo list's `load_from_disk` function now uses `add_todo` as
well, which ensures that newly created and restored todos all share the
same properties and signal connections.