Allow users to choose their store_path themselves
A form to do so will be shown when starting the app for the first time.
Once the value is set, it's stored in the settings.cfg and the form will
be skipped on subsequent starts of the app (unless the directory gets
deleted, in which case it will be shown again).

In the future, it should be possible to change this path from the (not
yet implemented) settings menu as well.