Make creating a successor to-do the default
When editing a to-do, inserting any non-whitespace text and pressing
KEY_ENTER, a successor to-do (with active edit mode) will now be created
below it by default. To create it above it instead, also hold shift.
Leaving the editing mode via clicking outside won't trigger it.

Also, introduces a placeholder text ("Enter to-do...") to the edit field
– and makes the necessary color additions to the theme to keep it easily
readable in both the dark and light color mode.

Lastly, modifies the `add_todo` function to also allow creating debug
todos with it (by passing an optional boolean `is_debug_todo` to it,
which defaults to false). This ensures they share the same signal
connections and modifications as regular todos.