Implement a setting to change the "Today" position
Currently there are three options: LEFTMOST (which is identical to the
old behavior), SECOND_PLACE (which will place the current day in the
second section and allow a glimpse at yesterday) and CENTERED (which
will place the current day at the center column).

Right now there is no way to change this value in the app itself.