Fix a wait_time bug in the day change timer
Setting the `wait_time` variable of an already running timer won't
actually change its wait time! Instead, the value must be passed as an
(optional) argument to its `start` method. Therefore, the `new day`
signal was emitted too often - which was more notable the closer to
midnight the app was launched, since the initial (and later reused)
`wait_time` was lower and the signal got emitted more frequently.