Add custom tooltips
Introduces a new Tooltip class, which is a simple node that will spawn a
PanelContainer when the mouse enters its parent, and frees it again when
the mouse exits the parent again. The tooltip will be placed on the left
or the right of the parent, whereever there's more window space. It will
be vertically centered relative to the parent, unless that would put the
tooltip (or parts of it) outside the window, then it gets clamped.

All this was necessary since Godot's default tooltips cannot be placed
freely. And while there is a virtual `_make_custom_tooltip` method to
customize them, this still does not allow free positioning because it
will be added as a child of the PopupPanel representing the tooltip.