Assign the calendar widget a minimum width
Previously, the calendar widget would resize depending on the length of
the month name. Now the minimum width is chosen in a way that it can
hold the longest month name (september) plus a bit of extra space. The
size flags of the weekday names and day buttons were changed so they use
all the available space.