Add configurable week start to the calendar widget
Introduces a new setting `start_week_on_monday`, which if `false` will
start new weeks on Saturday, else (as before) on Monday. The setting can
be toggled on and off in the settings panel and persists on restart.

Also increases the alpha value for DayButton's without any to-dos to
make them more readable, especially in the dark theme.