Initialize settings in `_ready`, not every opening
This also fixes a bug that would cause the settings panel to shrink a
few pixels on repeated openings, due to the OptionButton being cleared
and re-filled again while `fit_to_longest_item` was active.

Also changes the order of operations in `close_overlay`: First the
components are hidden, then the overlay itself. This avoids emitting
the `visibility_changed` signal of the components two times.