Fix a bug releasing focus on scroll events
The `main_window.gd` script contains a section to release the currently
focused element when the background of the window is clicked. But since
Godot considers scrolling the mouse wheel an `InputEventMouseButton` the
code also triggered when scrolling while editing a to-do item. The new
version will only release focus on `MOUSE_BUTTON_LEFT` events now.