Store & restore the folding state of headings
Folded headings are prefixed with "> " in the stored textfile, unfolded
ones with "v ". To preserve backwards compatibility, the prefix "# " is
considered equal to "v " and treated similarly.

Due to how the folding is currently implemented, the `folded` signal has
to be emitted with a one frame delay to allow `todo_list.gd` to recreate
all to-dos first, before folding the heading.