Hide CalendarWidget when clicking outside of it
... unless the click is already handled by an UI element there that is
*not* part of the `todo_list` scene and does *not* equal the background
of the main window (in which case it's handled by `unhandled_input`).

This admittedly feels quite hacky for how little it accomplishes, but it
was the only way I could get *exactly* the behavior that I wanted. With
this it's now possible to change the current day with the CalendarWidget
open, although the widget is not (yet) updating properly.