Fix tooltip not properly hiding on dragging to-do
Previously, the "Move To-Do" tooltip was not hiding properly when
dragging a to-do. It would only hide after dropping the item again.
Simply moving the line `get_node("Tooltip").hide_tooltip()` higher up in
the `_get_drag_data` function didn't solve this, since the `queue_free`
call for the `_tooltip_panel` wouldn't complete before duplicating the
item for the drag preview. So in addition I'm now calling `remove_child`
before the free, which implicitly also makes the tooltip disappear.