Always use LineEdit to render an item's text
That is, instead of switching back and forth between a Label and the
LineEdit node, remove the Label and just focus/unfocus the LineEdit.

This comes with a few limitations, though:
- It's no longer (easily) possible to render headings as uppercase
  letters because that would permanently alter the text.
- Unlike Labels, LineEdits do not support adding an ellipsis if the
  contained text is longer than the space available.

I'll likely ignore the first issue (users might not wish to capitalize
their headings after all, and if they do they can still do so manually).
However, the second issue probably needs to be adressed sooner or later.
It's just a visual problem, though, and fully functional nonetheless.