Modify UI with placeholders for planned features
Those features are:
- a "capture" panel, supposed to hold all to-dos that aren't yet (or
  again) assigned to a specific date,
- an "alarms" panel, supposed to track the time till/since to-dos
  that were specifically marked for tracking,
- a "search all" panel to search for to-dos via their text,
- a "search screen" panel which is supposed to only highlight matches
  among the to-dos currently visible on the screen.

Buttons for these planned features are grouped with the "Settings"
button and move to the bottom left. Since they all open some form or
panel, the overarching VBoxContainer is named "ExtraPanelsWidget".

The dark/light-mode toggle is moved below the view mode buttons, which
are moved to the bottom right side of the screen. Since they all affect
the appearance of the programm, they are named "ApperanceWidget".

Also, changes the behavior of the "Jump To Today" button in the sense
that it now always remains visible, but gets disabled instead of hidden
if the current day is already identical to today.