Implement screen search feature
Press the respective button in the bottom left of the window or use its
shortuct CTRL+F to open and focus the search panel. Type in a query and
any to-do that matches it, will be highlighted. Those that do not, will
get dimmed down to an alpha modulation value of 0.1. Quitting the search
(either by pressing ESC, CTRL+F or clicking anywhere else) will stop the
search and reset all highlights / changes in alpha modulation. However,
the search value is saved (as long as the programm isn't closed) and is
restored the next time the user opens the screen search panel.