Create a new "TopBar"
It contains:
- the Settings button,
- the SearchBar,
- and the ModeToggle.

This commit also adds a (placeholder) button to show the "Help" panel.

The computation of the minimum window size and the vertical padding at
the top and bottom end of the window were adapted accordingly.

All InputActions that change the current date were changed to require
the ALT key being pressed. This avoids triggering them by accident and
enables using them while the SearchBar is focused as well.

Also, fixes some missing changes to the theme on switching modes.