Fix a rare, but consistent segfault
The segfault would occur after triggering a tooltip on one side of the
window, then quickly moving the mouse to the other side of the window,
triggering the tooltip there, and then back to the first button.

I don't fully understand what happens here, but it seems to be some form
of race condition where the tooltip panel isn't freed quickly enough. So
instead of removing the panel via `remove_child`, I now simply hide it.
As explained in commit 877f844, this is required to hide the "Move To-Do"
tooltip properly when the dragging of a to-do item starts.