Add a (fold-out) side panel on the left side
This is supposed to be used for:
- the help panel
- the alarms panel
- the capture panel

The three respective buttons are functional now and unfold/fold the
panel when clicked. They also have shortcuts now: Press F1 to "show
help", F2 to "show alarms" and F3 to "show capture".
Consequently, "toggle_dark_mode" was remapped from F3 to F12.

The panel is scaled to match the size of a day column in the current
view mode, and should adapt to changes in the window size.