Fix empty tooltip shown after pressing "Set Alarm"
Previously, after pressing "Set Alarm", the button text would not only
change to "Remove Alarm", but also an empty tooltip would show up. The
reason was that `_spawn_panel` does *not* check if the tooltip text is
empty, and pressing a button would reset the `_hover_timer`.

Now, instead of only restarting the timer, the callback function calls
`show_tooltip`, which checks if all criteria for a tooltip are met.