Round a list's scroll offset to a multiple of 40
Previously, it was possible to edit the file a to-do list was stored in,
from outside of Habituary and change the "SCROLL:" value. That altered
value was then restored when booting up the app, potentially leading to
to-dos only being partially visible in the list.

Now, the list can still be edited outside of Habituary, but the value is
rounded to a multiple of 40 (the height of one entry in the list), which
ensures that all currently visible to-dos in the list are fully visible.