Add keyboard navigation between todo list items
It is now possible to press UP and DOWN to select the previous and next
item in a list (if there are any). Holding the key down will repeat the
action until it's released again or the beginning/end is reached.

Also, it is possible to press HOME or END to jump to the first or last
item in a given list immediately now.

Lastly, removes a (now outdated) comment from `day_panel.gd`, which has
been missed in the previous commit.