Generate theme(s) from code, pt. 1
This ultimately aims to facilitate creating and modifying themes in a
simple, traceable and efficient manner:

- Simple, because you can simply copy a similar block of code and adapt
  it to your needs (which, at least to my knowledge, cannot be easily
  achieved from within Godot's theme editor).
- Traceable, because reading the diff of a script is a lot easier than
  that of a resource file with cryptic references, at least to me.
- Efficient, because reusing the same e.g. StyleBox in multiple places
  should be a lot easier this way. (This is not done anywhere yet)

For now, this just adds an EditorScript that can be run to automatically
create a theme file in the `theme` directory. This file is not yet used
anywhere, nor is a complete replica of the current `default_theme` yet.