Fix side panel state not being restored correctly
Introduced through a change in 7df2fc4. Turns out, waiting for 2 frames
before emitting `side_panel_changed` was meant to wait until the entry
scene had been passed and switched over to the main scene. (This is not
an issue in the editor build, as it skips `entry_point.tscn` entirely.)
Thus, the presumed "fix" resulted in a bug that did *not* occur in the
editor, but in the exported build, and not restore the side bar.

Instead, we're  now simply making sure to call the connected callback
`_on_side_panel_changed` in `_ready` for all the relevant scenes.