Move `_create_style_box_flat` into `ConciseTheme`
The function wasn't changed  & can still be called like this:

  `button.normal = theme.create_style_box_flat(dict_of_properties)`

But now it's also possible to just pass the `dict_of_properties`:

  `button.normal = dict_of_properties`

and the `ConciseTheme` class will create the stylebox automatically.
Accordingly, `generator.gd` was updated to adapt that new way.

It is *not* possible to create `StyleBoxEmpty`, `StyleBoxTexture` or
`StyleBoxLine` nodes that way currently. In those cases the StyleBox
still needs to get created & passed along in `generator.gd`.

Also, adds some additional region blocks to `generator.gd`.