Add set_default_color method to ConciseTheme
This method will query all color properties from the default theme for
the theme type and set them to the value provided as an argument.

This makes it easier to theme a Control node with many different color
properties, e.g. a Button. The original way of setting colors remained
intact, so it is still possible to declare deviations from the default
value after having called set_default_color once.

The generator.gd script has been adapted to this, where appropriate.