Allow setting a "main style" for any theme type
This is similarly motivated as the "pin main style" option in Godot's
ThemeTypeEditor: Setting some properties for *all* StyleBox properties
of a certain theme type at once.

Unlike the "pin main style" option (which is not exposed to GDScript),
the resulting style boxes are completely independent of each other and
can still be changed afterwards to deviate from that norm.

This uses a new "update_style_box" function in ConciseTheme, which is
essentially a variant of the old "create_style_box_flat" function that
instead of creating a new StyleBoxFlat every time, will now re-use the
already existing style box and merely update the provided values.

The generator.gd script was updated accordingly.

Also, this undos the previous commit (2b38062) for now, as merging the
ConciseTheme with the default theme every time the generator is run, did
introduce a notable increase in the overall generation time, which (for
whatever reason) I didn't notice yesterday at all...