Re-use UIDs from the previous theme version
That is, if there's already a version of the generated theme on disk,
load that, check if it already contains a given type, and if so, apply
that type's UID to the same type in the newly generated theme, before
saving it to disk. This way, re-running the script without any changes
should no longer change the resulting theme files at all.