Apply new naming convention for scripts & scenes
(This is a test, and may or may not be reverted in the future)

With the way the project was structured previously, there was a lot of
redundancy since scenes usually resided inside their own directory with
just a scene and script file of the same name inside it, e.g.:

- dummy_scene/
  - dummy_scene.gd
  - dummy_scene.tscn

Now, these files are simply named "main":

- dummy_scene/
  - main.gd
  - main.tscn

This works best with the editor(!) setting "List Script Names As" set to
"Parent Directory And Name", which will *always* indicate the context of
a "main.gd" script, and not just when multiple are open at once.
There is no such setting for the scene tabs, although there it's usually
obvious by looking at the root node's name in the scene tree.

One drawback of this approach is that searching for scenes or scripts in
the FileSystem tab via their name becomes rather impractical, as these
no longer match the query and get filtered out of view. This is *not*
the case for the (fuzzy matching) "Quick Open" panels, though.