Adjust indentation levels after moving a to-do
Previously, it was possible to move an indented to-do to a different
location where that indentation level would be prohibited regularly.

Now, moving a to-do will re-indent it properly by calling the setter
function of its inentation_level property after the move again.
Also, the indentation_level property of all its (previous) sub items,
that is, to-dos that came after it in the list, and that had a higher
indentation level than the moved to-do, will be decreased by one.