Keep indentation level for predecessors/successors
That way, creating a to-do with <x> sub to-dos requires hitting TAB only
once, instead of once per sub to-do. This strikes a better balance, given
that the user needs to press SHIFT+TAB only once, too, to return to the
previous indentation level when they're done adding sub to-dos.