Adjust indentation level of sub items, too
Indenting a to-do with sub items (i.e. all subsequent to-dos that have
a higher indentation level then that to-do, until one doesn't) will now
also indent those sub item by the same amount.

Likewise, unindenting a to-do with sub item will unindent the sub items,
too, but first shifts the to-do to the last position of its scope (if it
isn't there already) to avoid turning its siblings into sub items of the
to-do by unindenting it at the old position, e.g.:

- ToDo A1
  - ToDo B1
    - ToDo C1
    - ToDo C2
  - ToDo B2
- ToDo A2

will become:

- ToDo A1
  - ToDo B2
- ToDo B1
  - ToDo C1
  - ToDo C2
- ToDo A2

instead of:

- ToDo A
- ToDo B1
  - ToDo C1
  - ToDo C2
  - ToDo B2
- ToDo A2

Also, increases the maximum indentation level from 3 to 5.