Add text color to the to-do formatting options
For now, there are five different colors, pre-defined in the Settings
autoload, following the "Aurora" part of the "Nord" color palette.

To-dos are initialized with no color (i.e. the default text color), but
can be set to a different color manually, by clicking (or scrolling) on
the newly added formatting option to the right of the italic toggle. If
the end of the color list is reached, it will wrap around.

Like the other formatting options, the text color is *only* applied to
the to-do itself, not its bookmark equivalent (if there is any).

The text color settings do persist across restarts of the app.