Limit searched files in _search_for_bookmarks
Previously, *all* files under the Settings.store_path were searched for
bookmarks – even if they clearly weren't to-do lists and the search was
guaranteed to not yield a result!

Now, the file name is first checked with a regular expressions to make
sure that it perfectly complies with Settings.date_format_save.