Save unsubmitted to-do text changes on tree exit
A previous commit (0ecf3da) introduced that unsubmitted to-do text is
saved when the app is closed. However, this did not match the behavior
when the list moved out of view, e.g. because the user changed the date.
This commit fixes that by connecting the code to the tree_exiting signal
(which is also emitted on a NOTIFICATION_WM_CLOSE_REQUEST).

Also, changes _start_debounce_timer, so that it checks if the timer is
still inside the tree before starting it, to avoid throwing errors in
case it's not, as the list is about to exit the tree. In those cases it
is not necessary to start the timer, just to set pending_save to true,
as the day panel will issue a save on exiting the tree anyway.

Lastly, renames the text_color property of todo items into text_color_id
for added clarity, and modifies its setter function, such that it only
emits the list_save_requested signal, if there was a value change.
The latter is useful, as _check_for_search_query_match merely retriggers
the setter to restore the text_color, without changing the value.