Set mouse inputs on the text color as handled
That is, using the mouse wheel to change a to-dos text color in a long
list, no longer scrolls the visible part of said list, too.

Furthermore, the text color's tooltip will be hidden as soon as any
valid inputs are detected (which is in line with the behavior of the
other formatting options).

Also, changes the _input function such that it's no longer necessary to
explicitly call accept_event() for each input caught. Instead, it will
be implicitly called at the end of the function. However, if the input
doesn't match any of the defined cases, the function will return before
that point, that is, the event won't be accepted then.