Fix EditingOptions auto-shrink on small list width
This replaces the previous, hardcoded pixel width threshold of the
EditingOptions with a value that is calculated during _ready.

This is achieved by calculating the combined minimum width of the (at
this point still unshrinked) EditingOptions during startup. For it to
be correct, the label of the bookmarks button needs to be temporarily
changed to the longer form ("Remove" instead of "Add"). Then, whenever
the size of the options reaches or falls below that value, it shrinks
automatically by hiding the FormatLabel and clipping button labels.