Auto-indent new to-dos based on their predecessor
New to-dos will no longer have an indentation level of zero, but start
at the level of their predecessor (if there is one) instead.

Furthermore, if that predecessor ends in a colon, add one level on top,
that is, automatically create the new to-do as its sub to-do.

Also, connects the resized signal of the EditingOptions to the function
_on_editing_options_resized from code, as the old way of connecting it
via the editor's GUI resulted in errors when opening the scene. I'm not
really sure why, presumably because the EditingOption's parent (that is,
the todo item) wasn't ready in time. The new way works better!