Add custom DragHoverTrigger node
This node shares almost all of its code with the previous drag border
script (which gets removed by this commit). However, it's modularized
better, which allows for reusing it in the ScrollButton scenes.

With that, it's now possible to scroll a to-do list up or down while
moving a to-do via drag and drop, simply by dragging it over the
respective ScrollButton and waiting long enough.

Furthermore, this commit includes some last-minute changes to the
todo_item script that I forgot to amend in yesterday's commit...