Fix EditingOptions being scaled wrong, initially
The issue arose as _on_editing_options_resized wasn't called in _ready.
So before the first resized signal was emitted, the wide version of the
EditingOptions would be used, even if there wasn't enough space. Which
then scaled up the parent to-do list to create the needed width.

Also, slightly adjusts the spacing of the buttons in the EditingOptions,
to make them look more uniform, and take up less overall space.