Auto-indent new to-dos added above a sub to-do
That is, given the following to-do list:

- Parent To-Do
  - Old Sub To-Do

A new to-do is added between those two, resulting in this:

- Parent To-Do
  - New Sub To-Do
  - Old Sub To-Do

Whereas, previously the result would have looked like this:

- Parent To-Do
- New Sub To-Do
  - Old Sub To-Do

That was especially dire as it was not possible to re-indent the "New
Sub To-Do" without further indenting the "Old Sub To-Do" as well – as
that would have been regarded as a sub to-do of the former now!

This is no longer the case now, as newly added to-do's, where self.text
is not yet set, will not trigger successor re-indents anymore.