Fix wrongly requesting a settings save on startup
This was caused by the setter functions for

- fade_ticked_off_todos
- and fade_non_today_dates

being explicitly called with their current value, which didn't result in
any changes to the variable value, but still started the debounce timer.

To prevent this from happening in the future, this commit adds a clause
to the beginning of all setter functions in autoloads/settings.gd, that
returns early if the variable is already set to the provided value.