Make the calendar's month and year clickable
Clicking them will bring up a PopupMenu (in the case of the month) or a
SpinBox (in the case of the year), which allow the user to quickly skip
to a specific month or year respectively.

Also, fixes a bug that would disable the "Today" button in the calendar
when the currently viewed month was equal to the current month, but the
year was not (which wasn't checked at all).

Furthermore, changes a bunch of color names in theme_generator.gd from
their global version (e.g. FROST_1) to their local one (e.g. primary_1).