Improve discoverability of the indentation feature
That is, if a to-do's text starts with "- ", and it hasn't yet reached
its maximum indentation level, strip the "- " from the text's beginning
and autotomatically indent it one level instead.

Furthermore, if a to-do's text is empty, but the user still presses the
backspace key, unindent it one level if possible.

Also, changes the maximum allowed indentation level from 5 back to 3,
and refactors the code for calcualting a to-dos individual maximum into
its own function get_maximum_indentation_level.

Lastly, adjusts the position of the Triangle node, when the indentation
level is changed, so it always points at the to-do's checkbox.