Add background panel for currently selected day
Restructures the DayPanel scene in such a way, that its root node now
actually is a PanelContainer, not just a VBox. This required changing
certain node paths, and reconnecting a few signals.

Technically, *all* DayPanels have a background now, but unless their
date matches the current_day they won't be visible.

They also won't be shown if the view_mode is set to 1.