Connect all signals from code only, pt. 1
This adds a _connect_signals function to most script files in the
project that's supposed to be called once in _ready and connect all
relevant signals to their callbacks.

Regions are used to separate global, parent and local signals (which
previously were often connected in the inspector).

This has the following advantages:
- One can see *all* connections in one place.
- Searching for a signal's name (via Ctrl+Shift+F) works.
- If a signal or callback is renamed/removed, there will be an error.

Note: This is an ongoing effort, not all files have been changed yet.