Simplify EntryPoint script
This was motivated by the desire to set the EntryPoint as the main scene
of the project, and get rid of its previous "editor"-override. Changing
the scene here takes one frame, which introduced a runtime difference of
one frame between the release and editor builds - now fixed!

I took the opportunity to refactor and clean up the script. It now uses
a _set_initial_state method (as introduced in the previous commit, for
the DayPanel) and shifts some of its logic to the settings autoload.