Undo new naming convention
As noted in the commit message of fdc72dc, this was merely a test, and
after using it for about a month, I fail to see any advantages...

During this month, I grew accustomed to using the "Quick Open" panels
(CTRL+SHIFT+O for scenes, CTRL+ALT+O for scripts). However, it's worth
noting that due to the way Godot's fuzzy matching works, these actually
work better with the *old* naming convention, too!

Take the following project structure, for example:

primary/
  secondary/
    - main.tscn
  - main.tscn

When looking to quick-open a scene where the path contains "primary", it
will return the longer match ("primary/secondary/main.tscn") first.

With the old naming convention, the structure would be:

primary/
  secondary/
    - secondary.tscn
  - primary.tscn

Here a fuzzy-search for "primary" would return "primary/primary.tscn"
first, as (unlike the other match) it contains "primary" twice.