Auto-save/load all settings.gd export properties
That is, it's no longer necessary to add a line to the save_to_disk and
load_from_disk methods, turning the variable into an export property is
enough. The section name is deferred from the export_group (if any).

This utilizes a new get_exported_properties function that is available
through the newly created autoload "Utils". It returns a dictionary of
export groups, each containing an array of all their export variables.
Export variables without a group are assigned to "no_group" instead.

Also, adds a debug message to the end of load_from_disk.

Furthermore, fixes two bugs with restoring the settings properly:
- In the dark_mode setter, the theme was previously applied to the
  current_scene (i.e. EntryPoint.tscn), not the tree root.
- The side_panel state was restored but then overriden by setting the
  initial state of the side panel in its own _ready function.