Change main scene of the project
Previously, the project would always load the entry_point scene first,
and then change to the main_window scene if the store_path was set.
Now, the project loads the main_scene right away, and only changes to
the entry_point scene if the store path is *not* set yet.

This optimizes the more likely case (of the store path being set), and
gets rid of the inherent one-frame delay of the old solution (which now
only occurs for loading the welcome screen, the much rarer case).

Also, renames entry_point into welcome_screen.