Set default clear color from within theme
That is, theme/generator.gd stores the neutral_1 color into a new export
variable called background_color which is applied when the ConciseTheme
is initialized (i.e. loaded into memory for the first time).

However, since the dark theme is set as the project's "custom" theme in
the project settings, Godot reuses its cached resource when switching to
to the dark variant of the theme, without the export being set and the
background color changing to the correct value...

To avoid this, autoloads/settings.gd now uses ResourceLoader.load() over
the regular load() function, which can be set to ignore the cache.