Make to-do text colors user-configurable
It's now possible to change the values of the set of five colors that
is cycled through when assigning a text color to a to-do item.

The setting persists across restarts of the app.

Also, changes the order of colors in the default set (to make green and
red easily accessible via only one press of the left/right mouse button,
roughly analog to left-/rightclicking a to-dos checkbox.