Add capture panel
This factors out a subset of the previous "DayPanel" scene, into its own
"ScrollabelTodoList" scene, adapting their scripts accordingly. That now
enables reusing it in the capture panel without specifying a date (still
required to create a DayPanel), but by directly setting the store_path.

To enable drag-and-drop between the list view and the capture panel, the
left margin of the drag overlay is now changed to zero while the capture
panel is open, and not only when the side panel is hidden entirely.

In the capture panel, the to-do's "Bookmark" button will be hidden for
now, as that would require a date. (This might change later on.)

Also, changes the background color for ToDoItem_Heading from neutral_3
to neutral_4, so it's distinguisable in the side panel as well.