Go to WelcomeScreen if store_path doesn't exist
Previously, the redirect would only happen when the store_path variable
was not *set*, but it was never checked if the path was actually valid!

Also, changes the default cursor shape of buttons in the WelcomeScreen
(to CURSOR_POINTING_HAND, or CURSOR_FORBIDDEN if disabled).

Lastly, slightly adjusts the generated themes to...
- make the button's hover color different from the background
- set the button's disabled color to a dimmed down version of the normal