Fix PopupMenu in CalendarWidget not hiding
... when moving the mouse out of the Month name button after clicking it
(and opening the PopupMenu) but *without* hovering over the menu itself,
that is, exiting the button to the left, right or top.

As soon as the PopupMenu has been hovered over with the mouse, it will
hide on mouse exit by default - no extra work required!

This also fixes an issue that would fully close the CalendarWidget, even
when clicking inside of it, while the PopupMenu is open.