Save & restore window state (maximized/fullscreen)
This adds new "AppState" settings `is_maximized` and `is_fullscreen`,
which can be set to a boolean value to indirectly change the window's
state. Like other settings this persists across restarts.

The implementation of this is a little weird, using getter and setter
functions which actually operate on the value of `get_window().mode`.
That way their final value (that gets saved to disk) is always in line
with the actual mode of the window, even if it was changed by directly
setting its value instead of doing it through its proxy-setting.