Fix wrong indentation level of newly added to-dos
When adding a to-do above an existing, indented to-do by clicking the
LineHighlight between that to-do and it's predecessor, the added to-do
did end up *not* being indented, unindenting its successor3.

It took me a while to figure out the cause, however, the fix is trivial:
Removing a superfluous signal connection to `_on_gui_input` that caused
the ToDoList's `add_todo` function to be called twice in a row.