Fix various issues with hiding ticked off to-dos
While the "Hide Ticked Off To-Dos" setting is active, those to-dos are
now hidden immediately when loading a DayPanel (either during startup or
due to changing the current date) or toggling the setting off and on.

When folding a heading and then unfolding it again, make sure to only
restore the visibility of to-dos that comply with the "Hide Ticked Off
To-Dos" setting, i.e. are unticked or have unticked sub items.

Introduces the new get_nearest_visible_predecessor|successor functions
to to_do_list.gd, which are now used to skip over hidden to-dos:
- "previous_todo" and "next_todo" input action
- selecting the next to-do after deleting one
- selecting the next to-do after hiding one
The last one has been newly added with this commit, and is only applied
when a to-do has been ticked off while its EditingOptions are shown.

Similarly, `find_item_pos` and `show_line_highlight` in todo_list.gd now
skip over non visible nodes, too. Otherwise the line could appear in the
wrong place or to-dos would have been added at false positions.

The fade out effect is now only applied to a to-do's CheckBox and
Content. That is, its EditingOptions are unaffected.

Toggling a to-do's state via the "toggle_todo" or "cancel_todo" input
actions is only possible when the to-do is not a heading now.