Rework view mode autochange based on window width
The old approach did subtract a bunch of magic numbers from the window
width to calculate the width of the list view, and was prone to errors.
For example, it didn't take into account the (back then: non-existent)
side_panel yet, and stopped working correctly while it was open.

The new approach still needs one magic number (the minimum width of a
to-do list, which cannot be set as its minimum size directly as that,
due to `wrap_controls` being enabled, would stop the user from shrinking
the window below the width necessary to display 7 lists in a row), but,
that aside, should be a lot more stable. It connects to the list_view's
`resized` signal, and directly utilizes its `size.x` property.

As the new approach uses a separate settings variable `view_mode_cap`,
the previous view mode can now easily be restored when the window is
scaled up again. Unless the user manually changes the `view_mode`, this
will happen automatically once there's enough space to raise the cap.