Set TodoList's minimum width
Previously it had none and could become smaller than the minimum width
of a to-do item, which led to a visible shift when editing a to-do in
such a situation. Now this shouldn't happen anymore!

The new minimum width is set to 160, which is large enough to ensure
that the shrunken version of a to-do's EditingOptions still fits in.

Ideally that value would be determined at runtime, not hardcoded, but
that's more involved, as to-do lists don't necessarily contain items,
and one cannot simply use `get_combined_minimum_size` as that doesn't
take into account that the EditingOptions of a to-do can shrink.