Fix most of the remaining texture scaling issues
Replaces notification_dot and triangle_up textures with an SVG variant.
The PNGs of line and dotted_connector were not replaced, as I couldn't
figure out how to make them keep their size while tiling yet.

Also, fixes the SVG scale of reset.svg (which was missed in the previous
commit) and drag_indicator.svg (which was set to 2, not 3).

Furthermore, changes the `anti_aliasing_size` of StyleBoxes in the theme
from 1.0 to 0.33, which makes them look less blurry on triple scale, and
also fixes thin black gaps (e.g. between the LeftSidebarButton buttons
and the SidePanel, or the EditingOptions and the Triangle).