Add new "Cache" autoload
... with the goal of reducing redundant disk accesses, and keeping track
of changes made outside the app (once Habituary is focused again).

This is a work in progress, and not yet used anywhere (e.g. when loading
the bookmarks, or the to-do lists of the list view). However, it already
serves as a working proof of concept: the cache data is kept up to date!

Also, adds a `store_path_changed` signal to the "Settings" autoload.