Fold headings based on indentation where possible
That is, if a heading has any sub items, folding it will only fold away
those, not all items until the next heading (if there is any).
If a heading has no sub items, it falls back to the old way of folding.

Also, the code accounts for nested headings (a heading with at least one
sub item that's a heading as well) now, too. If the sub item heading was
folded when folding its parent, it will be restored as folded as soon as
the parent gets unfolded again. Unlike before, a todo will no longer get
folded when the folded heading has a higher indentation level.

Lastly, headings are no longer included in the item count of a heading
(which is displayed while the heading is folded).