Fix get_subordinate_items returning wrong nodes
After the previous commit, `get_subordinate_items` (which returns the
nodes to be moved when the user starts dragging a to-do's drag handle)
still followed the old logic, without accounting for indented sub items
and nested headings. In other words, there was a difference in the nodes
affected by *folding* and *moving* a heading.

This commit aligns the logic of `get_subordinate_items` with that of the
reworked `fold_heading` function, and then refactors the latter to use
the return value of the `get_subordinate_items` function as well.